+++ reduct_cli/config.py
@@ -25,10 +25,10 @@ def write_config(path: Path, config: Config):
     if not Path.exists(path):
         os.makedirs(path.parent, exist_ok=True)
     with open(path, "w", encoding="utf8") as config_file:
-        config_file.write(toml.dumps(config))
+        toml.dump(config.dict(), config_file)
 
 
 def read_config(path: Path) -> Config:
     """Read config from TOML file"""
     with open(path, "r", encoding="utf8") as config_file:
-        return toml.loads(config_file.read())
+        return Config.parse_obj(toml.load(config_file))