+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-02-19
+
 ### Added
 
 - Wildcard support for `--entries` option in `rcli export`
@@ -105,7 +107,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Init release
 
-[Unreleased]: https://github.com/reductstore/reduct-cli/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/reductstore/reduct-cli/compare/v0.7.0...HEAD
+
+[0.7.0]: https://github.com/reductstore/reduct-cli/compare/v0.6.0...v0.7.0
 
 [0.6.0]: https://github.com/reductstore/reduct-cli/compare/v0.5.0...v0.6.0